Fix availability typos and rename files
