+++ client/src/screens/FilterScreen.jsx
@@ -139,7 +139,7 @@ const FilterScreen = () => {
                       </div>
                       <div className="flex items-center justify-between">
                       <Link
-                        to={`/abailability/${specialist._id}`} 
+                        to={`/availability/${specialist._id}`}
                         className="rounded-lg bg-cyan-700 px-5 py-2.5 text-center text-sm font-medium text-white hover:bg-cyan-800 focus:outline-none focus:ring-4 focus:ring-cyan-300 dark:bg-cyan-600 dark:hover:bg-cyan-700 dark:focus:ring-cyan-800"
                       >
                           Schedule appointment
